use pmmp start scripts for now
