+++ start.cmd
@@ -1,5 +1,5 @@
 @echo off
-TITLE Apollo-Legacy server software for Minecraft: Pocket Edition
+TITLE PocketMine-MP server software for Minecraft: Pocket Edition
 cd /d %~dp0
 
 if exist bin\php\php.exe (
@@ -8,26 +8,21 @@ if exist bin\php\php.exe (
 ) else (
 	set PHP_BINARY=php
 )
-if exist Apollo.phar (
-	set POCKETMINE_FILE=Apollo.phar
+
+if exist PocketMine-MP.phar (
+	set POCKETMINE_FILE=PocketMine-MP.phar
 ) else (
 	if exist src\pocketmine\PocketMine.php (
 		set POCKETMINE_FILE=src\pocketmine\PocketMine.php
-		
-	)
-	if exist SpigotPE.phar (
-        echo "hey you are using a old version we are now Apollo!"
-		pause
-		exit 1
 	) else (
-		echo "Couldn't find a valid Apollo installation"
+		echo "Couldn't find a valid PocketMine-MP installation"
 		pause
 		exit 1
 	)
 )
 
 if exist bin\mintty.exe (
-	start "" bin\mintty.exe -o Columns=88 -o Rows=32 -o AllowBlinking=0 -o FontQuality=3 -o Font="Consolas" -o FontHeight=10 -o CursorType=0 -o CursorBlinks=1 -h error -t "SpigotPE" -i bin/pocketmine.ico -w max %PHP_BINARY% %POCKETMINE_FILE% --enable-ansi %*
+	start "" bin\mintty.exe -o Columns=88 -o Rows=32 -o AllowBlinking=0 -o FontQuality=3 -o Font="Consolas" -o FontHeight=10 -o CursorType=0 -o CursorBlinks=1 -h error -t "PocketMine-MP" -i bin/pocketmine.ico -w max %PHP_BINARY% %POCKETMINE_FILE% --enable-ansi %*
 ) else (
 	%PHP_BINARY% -c bin\php %POCKETMINE_FILE% %*
 )
